More cleanup for compiler warnings, strcpy() and strcat()
The OpenBSD 7.7 C compiler provides a different view on code warnings
than the one offered by gcc.  On vm37, this has thrown up a number
of cases where our code could be improved:
- dead code removal
- one wrong variable use
- places where it would be safer to use use pmsrtrncat() instead of
  strcat() and ditto for strcpy()
- places where review of our use of strcpy()/strcat() shows the code
  to be OK (usually an alloc of the right size followed by strcat() or
  strcpy()) and these can be exceptioned in scripts/cull-build-warniongs
  so they do not appear in the summary produced in Logs/pcp by Makepkgs

  sprintf -> pmsprintf
    qa/qt/qmc_dynamic/qmc_dynamic.cpp
    qa/qt/qmc_group/qmc_group.cpp

  dead code removal
    qa/src/arch_maxfd.c
    qa/src/archinst.c
    qa/src/chk_metric_types.c
    qa/src/exertz.c
    qa/src/getversion.c
    qa/src/interp_bug.c
    qa/src/interp_bug2.c
    qa/src/mark-bug.c
    qa/src/multithread4.c
    qa/src/pdu-gadget.c
    qa/src/template.c

  strcpy|strcat -> pmstrncpy|pmstrncat
    qa/src/pmconvscale.c
    src/pmdas/mmv/mmvdump.c
    src/pmdas/mmv/src/mmv.c
    src/pmproxy/src/opentelemetry.c

  wrong variable used in diagnostic
    src/libpcp_web/src/discover.c

  printf long vs long long type mismatch
    src/pmlogcheck/pass2.c

  logic errors (not a real problem as pmview is the only user of this class)
    src/libpcp_qed/src/qed_app.cpp

  make comments match code
    src/libpcp/src/strings.c

  exceptions
    scripts/cull-build-warnings

+++ src/pmlogcheck/pass2.c
@@ -235,7 +235,7 @@ next:
 	    ;
 	}
 	else {
-	    fprintf(stderr, "%s.meta[record %d]: Botch BAD type meta off=%ld type=%d\n", archname, nrec, offset, h.type);
+	    fprintf(stderr, "%s.meta[record %d]: Botch BAD type meta off=%lld type=%d\n", archname, nrec, (long long)offset, h.type);
 	}
 	__pmFseek(f, (long)h.len - sizeof(h) - nread, SEEK_CUR);
 	nrec++;
